requirements.txt
discord.py==1.5.1
discord-py-slash-command==1.0.8.1